Defined a Data Loader class and transformation pipeline
Transformation pipeline resizes images to a resolution of 224x224, converts images to Tensor, and normalizes it by using ImageNet standards.

The Data Loader class loads the images and their corresponding labels, returns the total number of samples in  the dataset via __len__(self), and gets a specific sample when called by the dataloader via __getitem__(self, idx).

A print statement has been added for validation, it will be removed in a later commit.